requirements.txt
flask
newspaper3k
textblob
nltk
lxml_html_clean
